+++ main/src/components/ActionBar/ActionBar.jsx
@@ -47,7 +47,7 @@ function ActionBar() {
   const config = configManager.getConfig();
   const query = configManager.getQueryWorkingCopyById(resource);
   const context = query.comunicaContext;
-  const sources = context.sources;
+  const sources = context?.sources || []; // in early calls, context might be undefined
 
   return (
     <Grid container direction="row" width={"100%"} rowSpacing={1}>
@@ -102,10 +102,10 @@ function ActionBar() {
                       <SourceAuthenticationIcon source={source} />
                     </TableCell>
                     <TableCell>
-                      <SourceFetchStatusIcon proxyUrl={config.httpProxy} context={context} source={source} />
+                      <SourceFetchStatusIcon proxyUrl={config.httpProxy || ""} context={context} source={source} />
                     </TableCell>
                     <TableCell>
-                      <SourceVerificationIcon proxyUrl={config.httpProxy} context={context} source={source} />
+                      <SourceVerificationIcon proxyUrl={config.httpProxy || ""} context={context} source={source} />
                     </TableCell>
                   </TableRow>
                 ))}