+++ main/src/components/ListResultTable/QueryResultList/QueryResultList.jsx
@@ -1,5 +1,5 @@
 import { Component } from "react";
-import { Loading, Datagrid, ListView, Title, useListContext, useResourceDefinition } from "react-admin";
+import { Datagrid, List, Title, useListContext, useResourceDefinition } from "react-admin";
 import ActionBar from "../../ActionBar/ActionBar";
 import GenericField from "../../../representationProvider/GenericField";
 import TableHeader from "./TableHeader/TableHeader";
@@ -23,20 +23,6 @@ function QueryResultList(props) {
   const { resource, variableValues, changeVariables, submitted } = props;
   const resourceDef = useResourceDefinition();
   const queryTitle = resourceDef?.options?.label;
-  const { data, isLoading } = useListContext(props);
-  let values = {};
-  if (!isLoading && data && data.length > 0) {
-    data.forEach((record) => {
-      Object.keys(record).forEach((variable) => {
-        if (!values[variable]) {
-          values[variable] = [];
-        }
-        values[variable] = values[variable].concat(record[variable]);
-      });
-    });
-    delete values.id;
-  }
-
   const config = configManager.getConfig();
   const query = configManager.getQueryWorkingCopyById(resource);
 
@@ -45,7 +31,6 @@ function QueryResultList(props) {
   // LOG console.log(`isLoading: ${isLoading}`);
 
   return (
-    isLoading ? <Loading loadingSecondary={"The page is loading. Just a moment please."} /> :
     <div style={{ paddingLeft: '20px', paddingRight: '10px' }}>
       <Title title={config.title} />
       <div style={{ display: 'flex', flexDirection: 'row' }}>
@@ -60,23 +45,19 @@ function QueryResultList(props) {
         }
       </>
       }
-      {Object.keys(values).length ? (
-        <ListView title=" " actions={<ActionBar />} {...props} >
-          <Datagrid header={<TableHeader query={query} />} bulkActionButtons={false}>
-            {Object.keys(values).map((key) => {
-              return (
-                <GenericField
-                  key={key}
-                  source={key}
-                  label={key.split("_")[0]}
-                />
-              );
-            })}
-          </Datagrid>
-        </ListView>
-      ) :
-        <NoValuesDisplay />
-      }
+      <List
+        {...props}
+        disableAuthentication={true} // needed to overrule the default, which is to force logging in
+        title=" "
+        actions={ <ActionBar />}
+        empty={<NoValuesDisplay />}
+        queryOptions={{
+          keepPreviousData: false,
+          meta: {
+            variableValues: variableValues
+          }}}>
+        <MyDatagrid {...props} query={query} />
+      </List>
     </div>
   );
 }
@@ -92,22 +73,51 @@ const Aside = (props) => {
   const { changeVariables } = props;
   return (
     <Box>
-      <Button variant="contained" onClick={changeVariables} startIcon={<IconProvider.TuneIcon/>} sx={{ margin: '10px' }}>Change Variables</Button>
+      <Button variant="contained" onClick={changeVariables} startIcon={<IconProvider.TuneIcon />} sx={{ margin: '10px' }}>Change Variables</Button>
     </Box>
-    
-  )
+  );
 }
 
 const NoValuesDisplay = () => {
   return (
-<div>
-    <Box display="flex" alignItems="center" sx={{ m: 3 }}>
-      <SvgIcon component={SearchOffIcon} />
-      <span>The result list is empty.</span>
-    </Box>
+    <div>
+      <Box display="flex" alignItems="center" sx={{ m: 3 }}>
+        <SvgIcon component={SearchOffIcon} />
+        <span>The result list is empty.</span>
+      </Box>
     </div>
+  );
+}
 
-  )
+const MyDatagrid = (props) => {
+  const { query } = props;
+  const { data, isLoading } = useListContext(props);
+  let values = {};
+  if (!isLoading && data && data.length > 0) {
+    data.forEach((record) => {
+      Object.keys(record).forEach((variable) => {
+        if (!values[variable]) {
+          values[variable] = [];
+        }
+        values[variable] = values[variable].concat(record[variable]);
+      });
+    });
+    delete values.id;
+  }
+
+  return (
+    <Datagrid header={<TableHeader query={query} />} bulkActionButtons={false}>
+      {Object.keys(values).map((key) => {
+        return (
+          <GenericField
+            key={key}
+            source={key}
+            label={key.split("_")[0]}
+          />
+        );
+      })}
+    </Datagrid>
+  );
 }
 
 export default QueryResultList;