+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Using new verification library v0.2.0 (#184).
+
 ### Fixed
 
 - Queries no longer executed twice (#176).