+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Using new verification library v0.2.0 (#184).
+- Refactoring, avoiding the usage of an internal, undocumented react-admin component (#185).
 
 ### Fixed
 
@@ -21,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Templated queries: visiting with variable values in url search parameters now works as expected (#183).
 - Solved console warning when hovering over source verification question mark (#128).
 - ASK query is working now (#22).
+- "Finished in" now shows effective query execution time (#130).
 
 ## [1.6.0] - 2025-03-17